Adds check for whether cartan matrices are equivalent.
As the cartan matrix is a gram matrix, and we only care about the hyperplanes
which the are normal to the vectors, multiplying any vector by -1 makes no
difference. This would multiply the row and column in the gram matrix by -1
simulateously. Hence we need to consider two matrices equivalent if you can
go from one to the other by simultaneously mutliplying rows and columns by
-1. This is the check in cartan_equiv.