Having problems with equivalence test and zeros.
Everyhting is fine with the equivalence check until it encounters a matrix which
contains zeros. As 0 == -0 and 0 == 0 there is no easy way to decide whether to
flip that row or not. Somehow we need to come back to this choice later on, when
more of the matrix is checked.

Currently the tests pass, but there are almost certainly cases that I have
missed with more zeros.